decorator in utils ?
