+++ bioptim/models/biorbd/biorbd_model.py
@@ -1,5 +1,4 @@
 from typing import Callable
-from functools import wraps
 
 import biorbd_casadi as biorbd
 import numpy as np
@@ -15,7 +14,7 @@ from bioptim.models.biorbd.external_forces import (
     ExternalForceSetTimeSeries,
     ExternalForceSetVariables,
 )
-from ..utils import _var_mapping, bounds_from_ranges
+from ..utils import _var_mapping, bounds_from_ranges, cache_function
 from ...limits.path_conditions import Bounds
 from ...misc.mapping import BiMapping, BiMappingList
 from ...misc.utils import check_version
@@ -94,37 +93,6 @@ class BiorbdModel:
         external_force_set.bind()
 
         return external_force_set
-
-    def cache_function(method):
-        """Decorator to cache CasADi functions automatically"""
-
-        def make_hashable(value):
-            """
-            Transforms non-hashable objects (dicts, and lists) into hashable objects (tuple)
-            """
-            if isinstance(value, list):
-                return tuple(make_hashable(v) for v in value)
-            elif isinstance(value, dict):
-                return tuple(sorted((k, make_hashable(v)) for k, v in value.items()))
-            elif isinstance(value, set):
-                return frozenset(make_hashable(v) for v in value)
-            return value
-
-        @wraps(method)
-        def wrapper(self, *args, **kwargs):
-            # Create a unique key based on the method name and arguments
-            key = (method.__name__, args, frozenset((k, make_hashable(v)) for k, v in kwargs.items()))
-            if key in self._cached_functions:
-                return self._cached_functions[key]
-
-            # Call the original function to create the CasADi function
-            casadi_fun = method(self, *args, **kwargs)
-
-            # Store in the cache
-            self._cached_functions[key] = casadi_fun
-            return casadi_fun
-
-        return wrapper
 
     @property
     def name(self) -> str:
