+++ bioptim/models/protocols/biomodel.py
@@ -1,10 +1,10 @@
 import numpy as np
 from typing import Protocol, Callable, Any
-from functools import wraps
 
 from casadi import MX, SX, Function
 from ...misc.mapping import BiMapping, BiMappingList
 from ...limits.path_conditions import Bounds
+from ..utils import cache_function
 
 
 class BioModel(Protocol):
@@ -13,37 +13,6 @@ class BioModel(Protocol):
     bioptim.
     As a reminder for developers: only necessary attributes and methods should appear here.
     """
-
-    def cache_function(method):
-        """Decorator to cache CasADi functions automatically"""
-
-        def make_hashable(value):
-            """
-            Transforms non-hashable objects (dicts, and lists) into hashable objects (tuple)
-            """
-            if isinstance(value, list):
-                return tuple(make_hashable(v) for v in value)
-            elif isinstance(value, dict):
-                return tuple(sorted((k, make_hashable(v)) for k, v in value.items()))
-            elif isinstance(value, set):
-                return frozenset(make_hashable(v) for v in value)
-            return value
-
-        @wraps(method)
-        def wrapper(self, *args, **kwargs):
-            # Create a unique key based on the method name and arguments
-            key = (method.__name__, args, frozenset((k, make_hashable(v)) for k, v in kwargs.items()))
-            if key in self._cached_functions:
-                return self._cached_functions[key]
-
-            # Call the original function to create the CasADi function
-            casadi_fun = method(self, *args, **kwargs)
-
-            # Store in the cache
-            self._cached_functions[key] = casadi_fun
-            return casadi_fun
-
-        return wrapper
 
     @property
     def name(self) -> str:
