+++ bioptim/models/protocols/holonomic_biomodel.py
@@ -1,44 +1,13 @@
 from typing import Protocol, Callable
-from functools import wraps
 
 from biorbd_casadi import GeneralizedCoordinates
 from casadi import MX, DM, Function
 from .biomodel import BioModel
 from ..holonomic_constraints import HolonomicConstraintsList
+from ..utils import cache_function
 
 
 class HolonomicBioModel(BioModel, Protocol):
-
-    def cache_function(method):
-        """Decorator to cache CasADi functions automatically"""
-
-        def make_hashable(value):
-            """
-            Transforms non-hashable objects (dicts, and lists) into hashable objects (tuple)
-            """
-            if isinstance(value, list):
-                return tuple(make_hashable(v) for v in value)
-            elif isinstance(value, dict):
-                return tuple(sorted((k, make_hashable(v)) for k, v in value.items()))
-            elif isinstance(value, set):
-                return frozenset(make_hashable(v) for v in value)
-            return value
-
-        @wraps(method)
-        def wrapper(self, *args, **kwargs):
-            # Create a unique key based on the method name and arguments
-            key = (method.__name__, args, frozenset((k, make_hashable(v)) for k, v in kwargs.items()))
-            if key in self._cached_functions:
-                return self._cached_functions[key]
-
-            # Call the original function to create the CasADi function
-            casadi_fun = method(self, *args, **kwargs)
-
-            # Store in the cache
-            self._cached_functions[key] = casadi_fun
-            return casadi_fun
-
-        return wrapper
 
     def set_holonomic_configuration(
         self,
